restored test in xgrid
